Add Authentication with the newest version of database

+++ Models/Location/Governorate.cs
@@ -4,28 +4,28 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace GraduationProject.Models.Location
 {
-    public class Governorate
-    {
-        public int Id { get; set; }
+	public class Governorate
+	{
+		public int Id { get; set; }
 
-        [Required, MaxLength(250), Column(TypeName = "varchar")]
-        public string Name { get; set; }
+		[Required, MaxLength(250), Column(TypeName = "varchar")]
+		public string Name { get; set; }
 
-        [Required, MaxLength(250)]
-        public string Name_AR { get; set; }
+		[Required, MaxLength(250)]
+		public string Name_AR { get; set; }
 
-        public byte[] Image { get; set; }
+		public byte[] Image { get; set; }
 
-        public ICollection<City> Cities { get; set; }
+		public ICollection<City> Cities { get; set; }
 
-        public Governorate()
-        {
+		public Governorate()
+		{
 
-        }
+		}
 
-        public Governorate(uint id)
-        {
-            Id = (int)id;
-        }
-    }
+		public Governorate(int id)
+		{
+			Id = id;
+		}
+	}
 }
